+++ eras/allegra/impl/src/Cardano/Ledger/Allegra/TxBody.hs
@@ -78,8 +78,6 @@ import Cardano.Ledger.Shelley.TxBody (
   ShelleyEraTxBody (..),
   ShelleyTxBody (..),
   Withdrawals (..),
-  totalTxDepositsShelley,
-  totalTxRefundsShelley,
  )
 import Cardano.Ledger.TreeDiff (ToExpr)
 import Cardano.Ledger.TxIn (TxIn (..))
@@ -384,9 +382,6 @@ instance Crypto c => ShelleyEraTxBody (AllegraEra c) where
   updateTxBodyL =
     lensMemoRawType atbrUpdate $ \txBodyRaw update -> txBodyRaw {atbrUpdate = update}
   {-# INLINEABLE updateTxBodyL #-}
-
-  getTotalDepositsTxBody = totalTxDepositsShelley
-  getTotalRefundsTxBody = totalTxRefundsShelley
 
 instance Crypto c => AllegraEraTxBody (AllegraEra c) where
   {-# SPECIALIZE instance AllegraEraTxBody (AllegraEra StandardCrypto) #-}